+++ examples/test.rb
@@ -1,9 +1,24 @@
 $LOAD_PATH.push File.expand_path('../../lib', __FILE__)
 require 'clin'
 
-shell = Clin::Shell.new
-a = []
-# a << shell.choose('Is it true 1?', %w(usa france germany italy uk), allow_initials: true)
-a << shell.choose('Is it true 1?', {usa: 'United states', france: 'France'}, allow_initials: true)
 
-puts a.to_s
+class Test
+
+  def foo
+    'foo'
+  end
+
+  private def bar
+    'bar'
+  end
+
+  def pub
+    'pub'
+  end
+end
+
+
+t = Test.new
+puts t.foo
+puts t.pub
+puts t.bar