+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.3.0
+Features:
+    - Added a shell class for any user interaction.
+    - Added `auto_option` to define specific options cleaner.(#5)
+
 ## 0.2.0
 Features:
     - Allow unknown options to be ignored and not raise Error(#1)